Aggregate Propcheck fixtures to external module
